feat: Implement comprehensive modules for Leaves, Payroll Loans, Recruitment, and Performance management, including API endpoints, application logic, database entities, and frontend components.

+++ Backend/HRMS/HRMS.Application/Features/Payroll/Loans/Queries/GetEmployeeInstallments/GetEmployeeInstallmentsQuery.cs
@@ -32,7 +32,7 @@ public class GetEmployeeInstallmentsQueryHandler : IRequestHandler<GetEmployeeIn
 
         var dtos = installments.Select(i => new LoanInstallmentDto
         {
-            InstallmentId = i.InstallmentId,
+            InstallmentId =(int)i.InstallmentId,
             LoanId = i.LoanId,
             EmployeeId = i.Loan.EmployeeId,
             EmployeeName = i.Loan.Employee.FullNameAr,
